% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{addAnalysesToRun}
\alias{addAnalysesToRun}
\title{add models to the analysesToRun list.}
\usage{
addAnalysesToRun(data, modelMd = modelMetadata)
}
\arguments{
\item{data}{a \code{data} object}

\item{modelMd}{a \code{modelMetadata} object to retrieve analyses from, by default the package's \code{modelMetadata}}
}
\value{
a \code{data} object
}
\description{
add models to the analysesToRun list.
}

