% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseExtractedAnnotations.R
\name{addNewDataSets}
\alias{addNewDataSets}
\title{Extract new dataSets from a data object that includes fullData}
\usage{
addNewDataSets(data, dataSets, modelMd = modelMetadata)
}
\arguments{
\item{data}{a data object that has a fullData object.}

\item{dataSets}{A vector of the types of periods (aka: dataSets) to extract for analysis. Possible values are: "action", "estimation", "release", "estMaxGrip", "gestMaxGrip", and "gestMove"}

\item{modelMd}{a modelMetadata object, default: the modelMetadata from the package}
}
\value{
a data object. Wellformedness of this object can be checked with \code{\link{checkData}}
}
\description{
Add new dataSets to a data object that includes fullData.
}

