% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFunctions.R
\name{ceilingFinder}
\alias{ceilingFinder}
\title{Processing function for finding means or medians within a specific band at the top of the range of a grip}
\usage{
ceilingFinder(data, percOcclusion = 0.05, bandWidth, bandType = "percent")
}
\arguments{
\item{data}{the data to process.}

\item{percOcclusion}{the percentage of occlusion that is acceptable (this is the upper bound, in percent.) Default: \code{0.05}, or only trials where there is less than 5\% of occlusion are processed.}

\item{bandWidth}{the width of the band at the top of the range}

\item{bandType}{a character, either "percentage" if the band given should be used as a percentage of the range (this is the default); or "mm" if the band given should be used as millimeters}
}
\value{
a dataframe with the data that has been processed, one line per observation
}
\description{
\code{ceilingFinder()} takes the dataSet and extracts the mean and median of grip only for the period during which the grip is within a band at the top of the range of the grip measurement for the whole period. The range can be specified as either a percentage of the range or in millimeters.
This function should only be used in \code{\link{processDataSet}}, because that has the code to group all
 of the data into the right groups based on subject, session, trial, condition, period, etc. This should be
 included in the processing section of \link{modelMetadata} called \code{processFunction}
}

