% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{checkData}
\alias{checkData}
\title{check the form of data object}
\usage{
checkData(data, modelMd)
}
\arguments{
\item{data}{a data object to be checked}

\item{modelMd}{a modelStructure to use (to confirm that the dataSets are in the modelStructure)}
}
\value{
if the data objects form is correct, the data object
}
\description{
check the form of data object
}

