% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetadata.R
\name{checkmodelMetadata}
\alias{checkmodelMetadata}
\title{Check that modelMetadata conforms to the standard}
\usage{
checkmodelMetadata(modelMd)
}
\arguments{
\item{modelMd}{a modelMetadata object}
}
\value{
the modelMetadata object that was checked (if it passes)
}
\description{
Check that modelMetadata conforms to the standard
}

