% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetadata.R
\name{checkVariablesToUse}
\alias{checkVariablesToUse}
\title{Check that modelMetadata has all the variables and explanations that are needed}
\usage{
checkVariablesToUse(modelMd)
}
\arguments{
\item{modelMd}{a modelMetadata object}
}
\value{
the modelMetadata object that was checked (if it passes)
}
\description{
Check that modelMetadata has all the variables and explanations that are needed
}

