% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{displayAnalysesToRun}
\alias{displayAnalysesToRun}
\title{display the models that are set to run for all data}
\usage{
displayAnalysesToRun(data)
}
\arguments{
\item{data}{a \code{data} object}
}
\value{
named character vectors of models to run.
}
\description{
display the models that are set to run for all data
}

