% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataTools.R
\name{displayAnalysesToRunOne}
\alias{displayAnalysesToRunOne}
\title{display the models that are set to run for one dataSet}
\usage{
displayAnalysesToRunOne(name, dataSet)
}
\arguments{
\item{name}{the name of the dataSet}

\item{dataSet}{a dataSet}
}
\value{
named character vectors of models to run.
}
\description{
\bold{Warning:} the name must match the dataSet
}

