% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMocapDataFromAnnotations.R
\name{extractMocapDataFromAnnotations}
\alias{extractMocapDataFromAnnotations}
\title{Extract (and check) annotations from elan files.}
\usage{
extractMocapDataFromAnnotations(files, destDir)
}
\arguments{
\item{files}{A vector of elan files}

\item{destDir}{The destination directory where the extracted csvs should be saved}
}
\value{
Nothing, not useful currently.

The \code{files} option can be a single (fully specified) elan file, a vector of fully specific elan files, or it can be a path with wildcards (e.g. \code{./GRI_0??/GRI_0??-SESSION_0??-TRIAL_0??.eaf}) that matches at least one, but potentially many elan files
}
\description{
Takes in a vector of elan files (*.eaf) and extracts the annotations from them. Additionally it preforms some checking to make sure that the annotations are in the format we expect. This won't catch all possible annotation errors, but should catch many of the most common ones.
}

