% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{findTheBestModel}
\alias{findTheBestModel}
\title{Pick one model as the best}
\usage{
findTheBestModel(models, last = FALSE)
}
\arguments{
\item{models}{A list of models, as generated by \code{\link{fitModels}}}

\item{last}{A logical, should the last model be picked? By default this is \code{FALSE}, which means the first model will be picked.}
}
\value{
Returns a list of length two, the first, named \code{chosenModel} is the model chosen, \code{allModels} has all of the models, and is a copy of \code{models}
}
\description{
\code{findTheBestModel} finds the first (by default) model in a list that does not have convergence errors. As the formula lists are configured now, this will return the most complex model (that does not have convergence errors). This does not run any model comparison, it just finds the first (or last if specified with option \code{last=TRUE})
}

