% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{fitModels}
\alias{fitModels}
\title{Fit the right models based on the name of the dataset being passed}
\usage{
fitModels(type, data)
}
\arguments{
\item{type}{this is the type of data that is going to be analyzed from \code{data}. Examples: "action", "estimation" The name of this option determines which model will be fit.}

\item{data}{a data object as generated by \code{\link{readExtractedMocapData}}}

\item{additionalModelTypes}{additional types of models that are not currently programed (this may change.)}
}
\value{
a list of fit models
}
\description{
Fit the right models based on the name of the dataset being passed
}

