% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elanUtils.R
\name{fixPaths}
\alias{fixPaths}
\title{Fix the paths of elan files}
\usage{
fixPaths(files)
}
\arguments{
\item{files}{A vector of video files}
}
\value{
Nothing, not useful currently.
}
\description{
This function takes in a vector of elan files and it searches the path where the files are for video, audio, csv, and tsconf files to link to if the links do not currently work.
}
\details{
It uses a search that matches the filename and as much as the path as it can, but the search is limited to the folder structure for the project. In other words, it will only search in the folders AUDIO, Clipped Videos, mocapCSVs, (which must be located in the parent of the parent folder where the elan files are located) additionally, it searches in the folder containing the elan file for tsconf file links.
}

