% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elanUtils.R
\name{installPyelan}
\alias{installPyelan}
\title{Install pyelan from source}
\usage{
installPyelan()
}
\value{
Nothing, not useful currently.
}
\description{
By default, installing this r package directly from github will not install the submodule dependency of pyelan.
}

