% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grip.R
\name{mainFunc}
\alias{mainFunc}
\title{Main function for ingesting motion capture data.}
\usage{
mainFunc(files)
}
\arguments{
\item{files}{A vector of video files}
}
\value{
Unclear, but not useful currently.
}
\description{
This function takes in a vector of video files, and writes out elan files with videos linked and synchronized motion capture data linked. It depends on the raw motion capture data, videos, and audio files being in set predetermined folders already.
}

