% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripClipping.R
\name{makeElanFiles}
\alias{makeElanFiles}
\title{Make (blank) elan files from motion capture, video, and audio data.}
\usage{
makeElanFiles(files)
}
\arguments{
\item{files}{A vector of video files}
}
\value{
Nothing, not useful currently.
}
\description{
This function takes in a vector of video files, and writes out elan files with videos linked and synchronized motion capture data linked. It depends on the raw motion capture data, videos, and audio files being in set predetermined folders already.
}

