% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFunctions.R
\name{maxGripFinder}
\alias{maxGripFinder}
\title{Processing function for finding maximum grips}
\usage{
maxGripFinder(data, percOcclusion = 0.05)
}
\arguments{
\item{data}{the data to process.}

\item{percOcclusion}{the percentage of occlusion that is acceptable (this is the upper bound, in percent.) Default: \code{0.05}, or only trials where there is less than 5\% of occlusion are processed.}
}
\value{
a dataframe with the data that has been processed, one line per observation
}
\description{
\code{maxGripFinder()} takes the period and extracts the maximum value of grip for the whole period.
This function should only be used in \code{\link{processDataSet}}, because that has the code to group all
 of the data into the right groups based on subject, session, trial, condition, period, etc. This should be
 included in the processing section of \link{modelMetadata} called \code{processFunction}
}

