% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processingFunctions.R
\name{meanMedianSubsetFinder}
\alias{meanMedianSubsetFinder}
\title{Processing function for finding means or medians of a subset of the period}
\usage{
meanMedianSubsetFinder(data, percOcclusion = 0.05, start, stop,
  timeType = "msecs")
}
\arguments{
\item{data}{the data to process.}

\item{percOcclusion}{the percentage of occlusion that is acceptable (this is the upper bound, in percent.) Default: \code{0.05}, or only trials where there is less than 5\% of occlusion are processed.}

\item{start}{when to start the subset, either in percentage (default) or milliseconds}

\item{stop}{when to stop the subset, either in percentage (default) or milliseconds}

\item{timeType}{a character, either "percentage" if the times given should be used as percentages (this is thedefault); or "msecs" if the times given should be used as miliseconds}
}
\value{
a dataframe with the data that has been processed, one line per observation
}
\description{
\code{meanMedianSubsetFinder()} takes the dataSet and extracts the mean and median of grip for a subset of the dataSet.
This function should only be used in \code{\link{processDataSet}}, because that has the code to group all
 of the data into the right groups based on subject, session, trial, condition, period, etc. This should be
 included in the processing section of \link{modelMetadata} called \code{processFunction}
}

