% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{modelAllData}
\alias{modelAllData}
\title{Pick one model as the bist}
\usage{
modelAllData(data, ...)
}
\arguments{
\item{data}{A list of data, as generated by \code{\link{readExtractedMocapData}}}

\item{...}{options to pass to \code{\link{fitModels}} (or \code{\link{findTheBestModel}}, if reprogrammed adjusted)}
}
\value{
Returns a list very similar to the \code{data} that also includes a \code{models} list with each set of data within \code{data} (except for \code{allData}, if it is present.)
}
\description{
takes data, and returns an object that includes the originional data as well as fit models.
}

