% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetadata.R
\docType{data}
\name{modelMetadata}
\alias{modelMetadata}
\title{model analysis structure, and explanation information}
\format{An object of class \code{list} of length 4.}
\usage{
modelMetadata
}
\description{
This is read from modelMetadata.json, and different structures can be read by \code{readmodelMetadata()}
}
\keyword{datasets}

