% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelStructure.R
\docType{data}
\name{modelStructure}
\alias{modelStructure}
\title{model analysis structure, and explanation information}
\format{An object of class \code{list} of length 3.}
\usage{
modelStructure
}
\description{
This is read from modelStructure.json, and different structures can be read by \code{readModelStructure()}
}
\keyword{datasets}

