% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseExtractedAnnotations.R
\name{processDataSet}
\alias{processDataSet}
\title{takes a dataSet and data, and gives back a list with the extracted data in named lists.}
\usage{
processDataSet(dataSet, data, modelMd = modelMetadata)
}
\arguments{
\item{dataSet}{character the name of the dataSet}

\item{data}{a data object for the data to be retrieved from}

\item{modelMd}{a modelMetadata object, default: the modelMetadata from the package}
}
\value{
the extracted data from the dataSet
}
\description{
takes a dataSet and data, and gives back a list with the extracted data in named lists.
}

