% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseExtractedAnnotations.R
\name{readExportedMocapData}
\alias{readExportedMocapData}
\title{Read extracted motion capture data}
\usage{
readExportedMocapData(path)
}
\arguments{
\item{path}{Directory containing motion catpure csv files that were extracted with the \code{\link{extractAnnotations}} function.}
}
\value{
Not sure yet.
}
\description{
Reads in extracted motion capture data from a directory.
}

