% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetadata.R
\name{readModelMetadata}
\alias{readModelMetadata}
\title{reads model metadata from a specially formated json file}
\usage{
readModelMetadata(file)
}
\arguments{
\item{file}{a path to a json file that includes modelMetadata}
}
\value{
a modelMetadata object
}
\description{
This is useful if you want to use different types of models/analyses or dataSets. Pass the modelMetadata object that is returned form this object to functions that take it as an argument (e.g. \code{\link{readExtractedMocapData}}, \code{\link{makeReport}}) to override the package defaults.
Additionally, it preforms checks to make sure that the modelMetadata conforms to expected structure.
}

