% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMetadata.R
\name{writeModelMetadata}
\alias{writeModelMetadata}
\title{writes model metadata from a specially formated json file}
\usage{
writeModelMetadata(modelMd, path, overwrite = FALSE)
}
\arguments{
\item{modelMd}{a modelMetadata object}

\item{path}{a path to write a json file out}

\item{overwrite}{if the path exists, should it be overwritten?}
}
\value{
nothing
}
\description{
Writes the modelMetadata object to a json file. This is useful if you want to change the modelMetadata, but don't know what the structure should look like. As an additional precaution, the modelMetadata that is given is checked to make sure that it is valid before it is written.
}

