% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gripClipping.R
\name{extractAnnotations}
\alias{extractAnnotations}
\title{Main function for extracitng (and checking) annotations from elan files.}
\usage{
extractAnnotations(files)
}
\arguments{
\item{files}{A vector of video files}
}
\value{
Unclear, but not useful currently.
}
\description{
This function takes in a vector of elan files and extracts the annotations from them. Additionally it preforms some checking to makes ure that the annotations are in the format we expect. This won't catch all possible annotation errors, but should catch many of the most common ones.
}

