% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWriter.R
\name{makeReport}
\alias{makeReport}
\title{Make a full report from data}
\usage{
makeReport(data, reportPath = "./report.Rmd", title = "Grip Project Report",
  ...)
}
\arguments{
\item{data}{A list conforming to the structure of data for the project from \code{\link{extractMocapDataFromAnnotations}} (an example of this is the \code{\link{pureReplication}} object)}

\item{reportPath}{a string with the path and name of the report to be made (this should end in \code{.Rmd}), by default it is: \code{"./report.Rmd"}}

\item{title}{a string for the title of the report, the default is "Grip Project Report"}

\item{...}{options to pass to \code{\link{modelAllData}}, (e.g. \code{last=TRUE} if the last model that converged should be selected rather than the default (\code{last=FALSE}) first).}
}
\value{
An object that is a code of \code{data}, but with the models appened to each analysis present. This is useful if you need to inspect the models or use this for further analysis or reporting.
}
\description{
takes a data object, fits (and finds the best) model, and then produces a report based on the provided analysisSkeleton and narrative found in modelStructure.
}

