% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportWriter.R
\name{makeReport}
\alias{makeReport}
\title{Make a full report from data}
\usage{
makeReport(data, reportPath = "./report", title = "Grip Project Report",
  refitModels = FALSE, modelMd = modelMetadata, ...)
}
\arguments{
\item{data}{A list conforming to the structure of data for the project from \code{\link{extractMocapDataFromAnnotations}} (an example of this is the \code{\link{pureReplication}} object). If this data object includes fit models, they will not be refit by default.}

\item{reportPath}{a string with the path and name of the report to be made, by default it is: \code{"./report"} The final report will have this name (and path), and the extension \code{.html} Any extensions on the \code{reportPath} (like \code{.Rmd} or \code{.html}) will be ignored.}

\item{title}{a string for the title of the report, the default is "Grip Project Report"}

\item{refitModels}{logical, if there are models that are already fit, should they be refit? Default: \code{FALSE}}

\item{modelMd}{a modelMetadata object. If you have loaded a different modelMetadata object that you would like to use, place it here. The default is `modelMetadata` which comes with the package.}

\item{...}{options to pass to \code{\link{modelAllData}}, (e.g. \code{last=TRUE} if the last model that converged should be selected rather than the default (\code{last=FALSE}) first).}
}
\value{
Returns a data object (with all of the same data as the data object given in the \code{data} argument) with all of the models that were fit included.
}
\description{
takes a data object, fits (and finds the best) model, and then produces a report based on the provided analysisSkeleton and narrative found in modelMetadata. This will write two files: a polished html-based report (with the extension \code{.html}) and an intermediary R Markdown file that this html report is made from (with the extension \code{.Rmd}). For the most part, the R Markdown file can be ignored, the polished html report is the final product.
}

