% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.R
\name{modelAllData}
\alias{modelAllData}
\title{Model all data in a data object}
\usage{
modelAllData(data, refitModels = FALSE, modelMd = modelMetadata, ...)
}
\arguments{
\item{data}{A data object, as generated by \code{\link{readExtractedMocapData}}}

\item{refitModels}{logical, if there are models that are already fit, should they be refit? Default: \code{FALSE}}

\item{modelMd}{a modelMetadata object. If you have loaded a different modelMetadata object that you would like to use, place it here. The default is `modelMetadata` which comes with the package.}

\item{...}{options to pass to \code{\link{fitModels}} (or \code{\link{findTheBestModel}}, if reprogrammed adjusted)}
}
\value{
Returns a data object (with all of the same data as the data object given in the \code{data} argument) with all of the models that were fit included.
}
\description{
takes data, and returns an object that includes the original data as well as fit models.
}

