% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseExtractedAnnotations.R
\name{readExtractedMocapData}
\alias{readExtractedMocapData}
\title{Read extracted motion capture data}
\usage{
readExtractedMocapData(path, types = c("action", "estimation"))
}
\arguments{
\item{path}{Directory containing motion catpure csv files that were extracted with the \code{\link{extractMocapDataFromAnnotations}} function.}

\item{types}{A vector of the types of periods to extract. Default: c("action", "estimation") Possible values are: "action", "estimation", "release", "estMaxGrip"}
}
\value{
Not sure yet.
}
\description{
Reads in extracted motion capture data from a directory.
}

